023-[update] refactor code
