+++ backend/src/main/resources/application-prod.properties
@@ -8,4 +8,4 @@ spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.Ph
 server.error.include-stacktrace=on_param
 server.port=8080
 
-host=intproj24.sit.kmutt.ac.th/ssi4
+host=intproj24.sit.kmutt.ac.th
