+++ backend/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=backend
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/itbms?useUnicode=true&characterEncoding=utf8&characterSetResults=utf8&serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=02122547
+spring.datasource.password=345671
 spring.jpa.defer-datasource-initialization=true
 spring.data.jdbc.dialect=mysql
 spring.jpa.open-in-view=false
